% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difftime_office_hours.R
\name{dates_span_fun}
\alias{dates_span_fun}
\title{Create sequence of days in span
Creates a sequence with dates from \code{started} to \code{ended}, indcluding first + last day. Not vectorized.}
\usage{
dates_span_fun(started, ended)
}
\arguments{
\item{started}{An atomic vector of POSIX time for start time}

\item{ended}{An atomic vector of POSIX time for end time}
}
\value{
a numeric vector of number of working days in period
}
\description{
Create sequence of days in span
Creates a sequence with dates from \code{started} to \code{ended}, indcluding first + last day. Not vectorized.
}
\keyword{internal}

