% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difftime_office_hours.R
\name{difftime_office_hours}
\alias{difftime_office_hours}
\title{Calculate number of office hours between time stamps}
\usage{
difftime_office_hours(started, ended, working_hours = c(8, 16))
}
\arguments{
\item{started}{Start time of period, as POSIX}

\item{ended}{End time of period, as POSIX}

\item{working_hours}{Vector of lenght 2, start and end of office day in hours. Default c(8,16)}
}
\value{
Number of office hours between time stamps
}
\description{
\code{difftime_office_hours} Calculate number of office hours between time stamps. Office hours is default mon - fri, 8:00 to 16:00
}
\details{
More info here later? Or remove...
}
\author{
Stefan Jansson, \email{janzzon@gmail.com}
}

