% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difftime_office_hours.R
\name{difftime_office_hours_no_NA}
\alias{difftime_office_hours_no_NA}
\title{Internal calculation of office hours}
\usage{
difftime_office_hours_no_NA(started, ended, working_hours)
}
\arguments{
\item{started}{Start time of period, as POSIX. NA-values is NOT allowed}

\item{ended}{End time of period, as POSIX. NA-values is NOT allowed}

\item{working_hours}{Vector of lenght 2, start and end of office day in hours. Default c(8,16)}
}
\value{
Number of office hours between time stamps
}
\description{
Internal calculation of office hours between time stamps
}
\details{
Don't call this funciton externally, use \code{difftime_office_hours} that validates input and handles NA-values
}
\keyword{internal}

