% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difftime_office_hours.R
\name{time_as_duration}
\alias{time_as_duration}
\alias{work_days_n}
\title{POSIX to duration since mid night}
\usage{
time_as_duration(x)

work_days_n(x)
}
\arguments{
\item{x}{a vector of objects of class POSIX time}

\item{x}{is a vector of dates in POSIX-format}
}
\value{
object of class duration

a numeric vector of number of working days in period
}
\description{
Number of working days
Calculates number of work days, currently just by week days mon-fri
}
\keyword{internal}

