% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difftime_office_hours.R
\name{time_as_duration}
\alias{time_as_duration}
\title{POSIX to duration since mid night}
\usage{
time_as_duration(x)
}
\arguments{
\item{x}{a vector of objects of class POSIX time}
}
\value{
object of class duration
}
\keyword{internal}

