% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difftime_office_hours.R
\name{work_days_atomic}
\alias{work_days_atomic}
\title{Number of working days, atomic
Calculates number of work days, currently just by week days mon-fri}
\usage{
work_days_atomic(x)
}
\arguments{
\item{x}{is a vector of dates in POSIX-format}
}
\value{
a numeric vector of number of working days in period
}
\description{
Number of working days, atomic
Calculates number of work days, currently just by week days mon-fri
}
\keyword{internal}

