% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difftime_office_hours.R
\name{work_days_n}
\alias{work_days_n}
\title{Number of working days, wrapper
Calls \code{work_days_atomic} that calculates number of work days}
\usage{
work_days_n(x)
}
\arguments{
\item{x}{is a list of vector of dates in POSIX-format}
}
\value{
a numeric vector of number of working days in period
}
\description{
Number of working days, wrapper
Calls \code{work_days_atomic} that calculates number of work days
}
\keyword{internal}

